Refactored sample application code / branch.json template
