+++ assets/branch.json
@@ -1,0 +1,4 @@
+{
+  "useTestInstance": true,
+  "enableLogging":true
+}
