+++ lib/src/constants.dart
@@ -1,6 +1,5 @@
 class AppConstants {
   AppConstants._internal();
-  static const PLUGIN_VERSION = "8.0.0";
   static const MESSAGE_CHANNEL = 'flutter_branch_sdk/message';
   static const EVENT_CHANNEL = 'flutter_branch_sdk/event';
 }
